+++ tests/mocks/romram_controller.h
@@ -8,7 +8,7 @@ namespace emulator::mocks
     class MockMemoryController : public emulator::memory::MemoryController
     {
     public:
-        MOCK_METHOD(void, set, (uint32_t address, uint8_t value), (override));
-        MOCK_METHOD(uint8_t, get, (uint32_t  address), (const, override));
+        MOCK_METHOD(void, set, (uint16_t address, uint8_t value), (override));
+        MOCK_METHOD(uint8_t, get, (uint16_t  address), (const, override));
     };
 }
